refactor(virtual-mcp): extract LayoutTabContent into its own file
index.tsx was 1690 lines with several concerns interleaved. LayoutTabContent
(the "Layout" settings section — main-view select, show-chat toggle, and
pinned-views list) only reads its own props (virtualMcpId, form, flushAndSave)
and its local UITool/PinnedView/ConnectionWithTools types, with no shared
closures over the rest of the file, so it's a clean byte-identical move.
